chore: split children spec in separate file
